make tests runnable
